actualizations in services midleware token and functions of router to main

+++ src/services/AdminService.ts
@@ -2,24 +2,41 @@ import { ref, Ref } from 'vue';
 import { getDecodedToken } from './authHelpers';
 import AdminRegData from '../interfaces/AdminRegData';
 import DecodedToken from '../interfaces/DecodedToken';
+import { isTokenValid } from './authHelpers'; 
+import { useRouter } from 'vue-router'; // Para redirigir
+import { returnToken } from './authHelpers';
 const url: string = 'http://127.0.0.1:8000/api/admin/';
 export default class AdminService {
     private admins: Ref<AdminRegData[]>;
-    private admin: Ref<AdminRegData | undefined>;
+    private router = useRouter();
     constructor() {
         this.admins = ref<AdminRegData[]>([]);
-        this.admin = ref<AdminRegData | undefined>(undefined);
     }
-
+    private routerToMain(){
+        this.router.push({ name: 'MainV' });
+    }
     // Getter para admins
     getAdmins(): Ref<AdminRegData[]> {
         return this.admins;
     }
  
-    // Método para obtener todos 
+    // Método para obtener todos (show)
     async fetchAll(): Promise<void> {
         try {
-            const response: Response = await fetch(url);
+            if (!isTokenValid()) {
+                this.routerToMain(); // Redirige al usuario a la página de inicio
+                throw new Error('Token vencido');
+            } 
+            const token: string = returnToken();
+            const response: Response = await fetch(url,{
+                headers: {
+                    'Authorization': `Bearer ${token}`
+                },}
+            );                     
+            if (response.status === 401) {
+                this.routerToMain(); // Redirige si el token es inválido desde el servidor
+                throw new Error('No autorizado. Token inválido o vencido');
+            }
             const json: AdminRegData[] = await response.json();
             this.admins.value = json;
         } catch (e) {
@@ -27,7 +44,7 @@ export default class AdminService {
         }
     }
 
-    // Método para registrar admin
+    // Método para registrar admin (store)
     async useRegisterService() {
         const isLoading: Ref<boolean> = ref(false);
         const error: Ref<string | null> = ref(null);
@@ -45,13 +62,11 @@ export default class AdminService {
                         'Content-Type': 'application/json',
                     },
                     body: JSON.stringify(userData),
-                });
-    
+                });                
                 if (!response.ok) {
                     const errorData = await response.json();
                     throw new Error(errorData.message || 'Error en el registro.');
-                }
-    
+                }    
                 success.value = true;
             } catch (err) {
                 error.value = err instanceof Error ? err.message : 'Error desconocido.';
@@ -68,12 +83,24 @@ export default class AdminService {
         };
     } 
    
-    //Método para eliminar un admin
+    //Método para eliminar un admin (destroy)
     async deleteAdmin(id: string): Promise<boolean | null> {
         try {
+            if (!isTokenValid()) {
+                this.routerToMain(); // Redirige al usuario a la página de inicio
+                throw new Error('Token vencido');
+            }
+            const token: string = returnToken();
             const response: Response = await fetch(`${url}/${id}`, {
                 method: 'DELETE',
+                headers: {
+                    'Authorization': `Bearer ${token}`
+                }
             });
+            if (response.status === 401) {
+                this.router.push({ name: 'MainV' }); // Redirige si el token es inválido desde el servidor
+                throw new Error('No autorizado. Token inválido o vencido');
+              }
 
             if (!response.ok) {
                 throw new Error(`Error deleting admin: ${response.status}`);
@@ -87,9 +114,13 @@ export default class AdminService {
         }
     }
 
-    // Método para actualizar un admin
+    // Método para actualizar un admin (updatePartial)
     async updateAdmin(updateData: Partial<AdminRegData>, id: string): Promise<AdminRegData | null> {
         try {
+            if (!isTokenValid()) {
+                this.routerToMain(); // Redirige al usuario a la página de inicio
+                throw new Error('Token vencido');
+            }
             const decodedToken: DecodedToken | null = getDecodedToken();
             let data: Partial<AdminRegData>;
             if (decodedToken !== null) {
@@ -98,13 +129,19 @@ export default class AdminService {
                 throw new Error('token decoded null');
             }
             console.log(`updated whidht id ${id} and data ${ JSON.stringify(data)}`)
+            const token: string = returnToken();
             const response: Response = await fetch(`${url}/${id}`, {
                 method: 'PATCH',
                 headers: {
                     'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${token}`
                 },
                 body: JSON.stringify(data),
             });
+            if (response.status === 401) {
+                this.routerToMain(); // Redirige si el token es inválido desde el servidor
+                throw new Error('No autorizado. Token inválido o vencido');
+            }
             if (!response.ok) {
                 throw new Error(`Error updating admin: ${response.status}`);
             }
